Move to tests/remotes.py
